+++ pytorch_hccl_tests/osu/collectives/osu_gather.py
@@ -6,7 +6,7 @@ import logging
 import torch
 import torch.distributed as dist
 
-from pytorch_hccl_tests.commons import dist_init, get_device, log_env_info
+from pytorch_hccl_tests.commons import dist_init, get_device, log_env_info, get_dtype
 from pytorch_hccl_tests.osu.options import Options
 from pytorch_hccl_tests.osu.osu_util_mpi import Utils
 from pytorch_hccl_tests.osu.parser import get_parser
@@ -19,7 +19,7 @@ def osu_gather(args):
     backend = args.backend
     rank = dist.get_rank()
     world_size = dist.get_world_size()
-    dtype = torch.float32
+    dtype = get_dtype(args.dtype)
     device = get_device(backend, rank)
     pg = None