+++ pytorch_hccl_tests/osu/p2p/osu_bw.py
@@ -8,6 +8,7 @@ import torch.distributed as dist
 from pytorch_hccl_tests.commons import (
     get_device,
     wait_all,
+    get_dtype,
 )
 from pytorch_hccl_tests.osu.options import Options
 from pytorch_hccl_tests.osu.osu_util_mpi import Utils
@@ -19,7 +20,7 @@ def osu_bw(args):
     backend = args.backend
     rank = dist.get_rank()
     world_size = dist.get_world_size()
-    dtype = torch.float32
+    dtype = get_dtype(args.dtype)
     device = get_device(backend, rank)
     pg = None