+++ pytorch_hccl_tests/osu/collectives/__init__.py
@@ -1,3 +1,5 @@
-from pytorch_hccl_tests.osu.collectives.osu_allreduce import osu_allreduce  # noqa
-from pytorch_hccl_tests.osu.collectives.osu_broadcast import osu_broadcast  # noqa
-from pytorch_hccl_tests.osu.collectives.osu_alltoall import osu_alltoall  # noqa
+from pytorch_hccl_tests.osu.collectives.osu_allreduce import allreduce  # noqa
+from pytorch_hccl_tests.osu.collectives.osu_broadcast import broadcast  # noqa
+from pytorch_hccl_tests.osu.collectives.osu_alltoall import alltoall  # noqa
+from pytorch_hccl_tests.osu.collectives.osu_allgather import allgather  # noqa
+from pytorch_hccl_tests.osu.collectives.osu_reduce_scatter import reducescatter  # noqa