+++ pytorch_hccl_tests/osu/p2p/__init__.py
@@ -1,4 +1,4 @@
-from pytorch_hccl_tests.osu.p2p.osu_latency import osu_latency  # noqa
-from pytorch_hccl_tests.osu.p2p.osu_bw import osu_bw  # noqa
-from pytorch_hccl_tests.osu.p2p.osu_bibw import osu_bibw  # noqa
-from pytorch_hccl_tests.osu.p2p.osu_multi_lat import osu_multi_lat  # noqa
+from pytorch_hccl_tests.osu.p2p.osu_latency import latency  # noqa
+from pytorch_hccl_tests.osu.p2p.osu_bw import bw  # noqa
+from pytorch_hccl_tests.osu.p2p.osu_bibw import bibw  # noqa
+from pytorch_hccl_tests.osu.p2p.osu_multi_lat import multi_lat  # noqa