+++ pytorch_hccl_tests/osu/p2p/osu_bw.py
@@ -16,7 +16,7 @@ from pytorch_hccl_tests.osu.osu_util_mpi import Utils
 logger = logging.getLogger(__name__)
 
 
-def osu_bw(args):
+def bw(args):
     backend = args.backend
     rank = dist.get_rank()
     world_size = dist.get_world_size()