+++ pytorch_hccl_tests/osu/startup/__init__.py
@@ -1,1 +1,1 @@
-from pytorch_hccl_tests.osu.startup.osu_hello import osu_hello  # noqa
+from pytorch_hccl_tests.osu.startup.osu_hello import hello  # noqa