+++ HISTORY.rst
@@ -6,6 +6,7 @@ History
 ------------------
 * Fix critical bug on setting up the NPU/HCCL environment
 * Introduce `dtype` as an argument parameter
+* Port `all_gather` and `reduce_scatter` collectives
 
 0.1.1 (2023-06-15)
 ------------------